Update Music + Crypto Bug Fix
Music was updated to use the lavalink.py custom source as it makes it easier to serve the streams/info and does it in slightly less code. Also, a bug stopping the /prices command to stop working has been fixed. The issue was a mis-index to the price of the crypto.