Upload Pre-Made Slot Reels
This is for the coming change that will change the slots command. Rather than slot reels being made on the spot, it will just randomly pick a slot reel based on if the winner has won or lost. This helps to save resources, especially for when multiple users are playing slots at the same time.